Implement CPDF_CrossRefTable

+++ core/fxcodec/codec/fx_codec_jpx_opj.cpp
@@ -567,12 +567,13 @@ void CJPX_Decoder::GetInfo(uint32_t* width,
 }
 
 bool CJPX_Decoder::Decode(uint8_t* dest_buf,
-                          uint32_t pitch,
+                          int pitch,
                           const std::vector<uint8_t>& offsets) {
   if (m_Image->comps[0].w != m_Image->x1 || m_Image->comps[0].h != m_Image->y1)
     return false;
 
-  if (pitch<(m_Image->comps[0].w * 8 * m_Image->numcomps + 31)>> 5 << 2) {
+  if (pitch<static_cast<int>(m_Image->comps[0].w * 8 * m_Image->numcomps + 31)>>
+      5 << 2) {
     return false;
   }
 
@@ -590,14 +591,14 @@ bool CJPX_Decoder::Decode(uint8_t* dest_buf,
       }
     }
   }
-  uint32_t width = m_Image->comps[0].w;
-  uint32_t height = m_Image->comps[0].h;
+  int width = m_Image->comps[0].w;
+  int height = m_Image->comps[0].h;
   for (uint32_t channel = 0; channel < m_Image->numcomps; ++channel) {
     uint8_t* pChannel = channel_bufs[channel];
     if (adjust_comps[channel] < 0) {
-      for (uint32_t row = 0; row < height; ++row) {
+      for (int row = 0; row < height; ++row) {
         uint8_t* pScanline = pChannel + row * pitch;
-        for (uint32_t col = 0; col < width; ++col) {
+        for (int col = 0; col < width; ++col) {
           uint8_t* pPixel = pScanline + col * m_Image->numcomps;
           if (!m_Image->comps[channel].data)
             continue;
@@ -614,9 +615,9 @@ bool CJPX_Decoder::Decode(uint8_t* dest_buf,
         }
       }
     } else {
-      for (uint32_t row = 0; row < height; ++row) {
+      for (int row = 0; row < height; ++row) {
         uint8_t* pScanline = pChannel + row * pitch;
-        for (uint32_t col = 0; col < width; ++col) {
+        for (int col = 0; col < width; ++col) {
           uint8_t* pPixel = pScanline + col * m_Image->numcomps;
           if (!m_Image->comps[channel].data)
             continue;
@@ -661,7 +662,7 @@ void CCodec_JpxModule::GetImageInfo(CJPX_Decoder* pDecoder,
 
 bool CCodec_JpxModule::Decode(CJPX_Decoder* pDecoder,
                               uint8_t* dest_data,
-                              uint32_t pitch,
+                              int pitch,
                               const std::vector<uint8_t>& offsets) {
   return pDecoder->Decode(dest_data, pitch, offsets);
 }
